add decorators to storybook for mantine, react-query, styled-components
