+++ client/.storybook/main.ts
@@ -12,14 +12,20 @@ function getAbsolutePath(value: string): any {
 const config: StorybookConfig = {
 	stories: ["../src/**/*.mdx", "../src/**/*.stories.@(js|jsx|mjs|ts|tsx)"],
 	addons: [
-        getAbsolutePath("@storybook/addon-essentials"),
-        getAbsolutePath("@chromatic-com/storybook"),
-        getAbsolutePath("@storybook/addon-interactions"),
-        getAbsolutePath("@storybook/addon-themes")
-    ],
+		getAbsolutePath("@storybook/addon-essentials"),
+		getAbsolutePath("@chromatic-com/storybook"),
+		getAbsolutePath("@storybook/addon-interactions"),
+		getAbsolutePath("@storybook/addon-themes")
+	],
 	framework: {
 		name: getAbsolutePath("@storybook/react-vite"),
 		options: {}
-	}
+	},
+	previewHead: (head) => `
+      ${head}
+      <style>
+			@import url("https://fonts.googleapis.com/css2?family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700;1,900&display=swap");
+		</style>
+   `
 };
 export default config;
